agrego el paquete excepciones y refactorizo
