+++ src/main/java/edu/fiuba/algo3/excepciones/EjercitoConUnaFichaNoPuedeAtacar.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.modelo;
+package edu.fiuba.algo3.excepciones;
 
 public class EjercitoConUnaFichaNoPuedeAtacar extends RuntimeException{
 }
