+++ src/main/java/edu/fiuba/algo3/excepciones/EjercitoNoPuedeTirarEsaCantidadDeDados.java
@@ -1,4 +1,4 @@
-package edu.fiuba.algo3.modelo;
+package edu.fiuba.algo3.excepciones;
 
 public class EjercitoNoPuedeTirarEsaCantidadDeDados extends RuntimeException{
 }
