+++ src/main/test/edu/fiuba/algo3/modelo/DadosTest.java
@@ -1,5 +1,6 @@
 package edu.fiuba.algo3.modelo;
 
+import edu.fiuba.algo3.excepciones.NoSePuedenCrearCeroDados;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
