+++ src/main/test/edu/fiuba/algo3/modelo/TableroTest.java
@@ -1,5 +1,6 @@
 package edu.fiuba.algo3.modelo;
 
+import edu.fiuba.algo3.excepciones.JugadorNoPoseePaisException;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
